added sendTX func to interact with Rinkeby
