+++ kycEndpoint/.vscode/settings.json
@@ -1,4 +1,4 @@
 {
-    "python.pythonPath": "${workspaceFolder}\\apienv\\Scripts\\python.exe",
+    "python.pythonPath": "C:\\Python37\\python.exe",
     "python.linting.pylintEnabled": false
 }
